create useHealthData hook
