+++ .expo/types/router.d.ts
@@ -6,7 +6,7 @@ export * from 'expo-router';
 declare module 'expo-router' {
   export namespace ExpoRouter {
     export interface __routes<T extends string = string> extends Record<string, unknown> {
-      StaticRoutes: `/` | `/_sitemap` | `/common/api/api` | `/common/api/pokemon-calls` | `/common/interface/api.interface`;
+      StaticRoutes: `/` | `/_sitemap` | `/common/api/api` | `/common/api/pokemon-calls` | `/common/interface/api.interface` | `/hooks/useHealthData`;
       DynamicRoutes: never;
       DynamicRouteTemplate: never;
     }
